Added check to display only non-own messages

+++ hack/user-page-messages.php
@@ -74,7 +74,8 @@ if (isset($_GET['username'])) {
                 });
 
                 if (response.ok) {
-                    loadMessages(parseInt('<?php echo $userData['id']; ?>'));
+                    loadMessages(parseInt('<?php echo $userData['id']; ?>'), 25);
+
                     messageTextarea.value = '';
                 } else {
                     alert('Failed to message');
@@ -89,7 +90,7 @@ if (isset($_GET['username'])) {
 
 
 
-    <script>
+    <!-- <script>
         async function loadMessages(recipientId, senderId) {
             const messagesContainer = document.getElementById('messagesContainer');
             try {
@@ -111,6 +112,74 @@ if (isset($_GET['username'])) {
 
                 if (Array.isArray(messagesData.success)) {
                     const messages = messagesData.success;
+                    console.log("messages", messages);
+
+                    const promises = messages.map(async message => {
+                        const senderId = message.sender_id;
+                        const userDataResponse = await fetch(`hack/messages/get_user_by_id.php?id=${senderId}`);
+                        const userData = await userDataResponse.json();
+                        console.log("userData", userData);
+                        return {
+                            ...userData,
+                            ...message
+                        };
+                    });
+
+                    Promise.all(promises).then(messagesWithUserData => {
+                        const messagesHTML = messagesWithUserData.map(message =>
+                            `<li class="message-li">
+                        <a class="message-a" href='index.php?page=user&username=${encodeURIComponent(message.name)}'>
+                            <img class="message-img" src='hack/${message.avatar}' alt='${message.name}'> 
+                            <div class="message-div">
+                                <div class="message-blk">
+                                    <p class="message-name">${message.name}</p>
+                                    <p class="message-a__text">${message.message_text}</p>
+                                </div>
+                                <button class="message-a__button" onclick="deleteMessage(${message.id}, event)">Delete</button>
+                            </div>
+                        </a>
+                    </li>`
+                        ).join('');
+                        messagesContainer.innerHTML = messagesHTML;
+                        console.log("messagesWithUserData", messagesWithUserData);
+                    });
+                } else {
+                    console.error('Response data is not in the expected format');
+                }
+            } catch (error) {
+                console.error('Error in fetch request', error);
+            }
+        }
+
+        loadMessages(parseInt(''), 25);
+    </script> -->
+
+
+
+    <script>
+        async function loadMessages(recipientId, senderId) {
+            const messagesContainer = document.getElementById('messagesContainer');
+            try {
+                const messagesResponse = await fetch('hack/messages/get_messages.php', {
+                    method: 'POST',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                    body: JSON.stringify({
+                        recipient_id: recipientId,
+                    }),
+                });
+
+                console.log("messagesResponse", messagesResponse)
+
+                if (!messagesResponse.ok) {
+                    throw new Error('Failed to fetch messages');
+                }
+
+                const messagesData = await messagesResponse.json();
+
+                if (Array.isArray(messagesData.success)) {
+                    const messages = messagesData.success;
 
                     // Отримати дані про користувача
                     const userDataResponse = await fetch('hack/messages/get_user_by_id.php', {
@@ -119,15 +188,19 @@ if (isset($_GET['username'])) {
                             'Content-Type': 'application/json',
                         },
                         body: JSON.stringify({
-                            sender_id: senderId
+                            id: senderId
                         }),
                     });
+
+                    console.log("userDataResponse", userDataResponse)
 
                     if (!userDataResponse.ok) {
                         throw new Error('Failed to fetch user info');
                     }
 
                     const userData = await userDataResponse.json();
+
+                    console.log("userData", userData)
 
                     // Об'єднати дані про користувача з кожним повідомленням
                     const messagesWithUserData = messages.map(message => {
@@ -138,22 +211,24 @@ if (isset($_GET['username'])) {
                     });
 
                     console.log("messagesWithUserData", messagesWithUserData)
-
+                    const loggedInUserId = parseInt('<?php echo $userData['id']; ?>');
                     // Відобразити повідомлення
-                    const messagesHTML = messagesWithUserData.map(message =>
-                        `<li class="message-li">
-                    <a class="message-a" href='index.php?page=user&username=${encodeURIComponent(message.success.name)}'>
-                        <img class="message-img" src='hack/${message.success.avatar}' alt='${message.success.name}'> 
-                        <div class="message-div">
-                            <div class="message-blk">
-                                <p class="message-name">${message.success.name}</p>
-                                <p class="message-a__text">${message.message_text}</p>
-                            </div>
-                            <button class="message-a__button" onclick="deleteMessage(${message.id}, event)">Delete</button>
-                        </div>
-                    </a>
-                </li>`
-                    ).join('');
+                    const messagesHTML = messagesWithUserData.map(message => {
+                        if (message.sender_id !== loggedInUserId) {
+                            return `<li class="message-li">
+                                        <a class="message-a" href='index.php?page=user&username=${encodeURIComponent(message.success.name)}'>
+                                            <img class="message-img" src='hack/${message.success.avatar}' alt='${message.success.name}'> 
+                                            <div class="message-div">
+                                                <div class="message-blk">
+                                                    <p class="message-name">${message.success.name}</p>
+                                                    <p class="message-a__text">${message.message_text}</p>
+                                                </div>
+                                                <button class="message-a__button" onclick="deleteMessage(${message.id}, event)">Delete</button>
+                                            </div>
+                                        </a>
+                                    </li>`
+                        }
+                    }).join('');
 
                     messagesContainer.innerHTML = messagesHTML;
                 } else {
@@ -164,7 +239,7 @@ if (isset($_GET['username'])) {
             }
         }
 
-        loadMessages(parseInt('<?php echo $userData['id']; ?>'), 25);
+        loadMessages(parseInt('<?php echo $userData['id']; ?>'), 26);
     </script>
 
 
@@ -185,7 +260,7 @@ if (isset($_GET['username'])) {
 
                 console.log('messageId', messageId);
 
-                loadMessages(parseInt('<?php echo $userData['id']; ?>'));
+                loadMessages(parseInt('<?php echo $userData['id']; ?>'), 25);
 
             } catch (error) {
                 console.error('Error:', error);
